requirements.txt
discord.py
r6sapi
